feat: add global inertia page props
