+++ resources/js/src/types/inertia.d.ts
@@ -1,0 +1,22 @@
+import '@inertiajs/core';
+
+type LangValue = string | { [key: string]: LangValue };
+export type LangObject = Record<string, LangValue>;
+
+declare module '@inertiajs/core' {
+  interface PageProps {
+    csrf_token: string;
+    locale: string;
+    lang: LangObject;
+    auth: {
+      user?: {
+        name?: string;
+        email?: string;
+        avatar?: string;
+        role?: string;
+        roles?: string[];
+        permissions?: string[];
+      };
+    };
+  }
+}
